♻refactor[#64]: 월 목표 validator 적용
