+++ src/main/java/com/sillim/recordit/goal/dto/response/MonthlyGoalListResponse.java
@@ -1,10 +1,10 @@
-package com.sillim.recordit.goal.controller.dto.response;
+package com.sillim.recordit.goal.dto.response;
 
 import com.sillim.recordit.goal.domain.MonthlyGoal;
 
 public record MonthlyGoalListResponse(Long id, String title, Boolean achieved) {
 
-	public static MonthlyGoalListResponse from(MonthlyGoal monthlyGoal) {
+	public static MonthlyGoalListResponse from(final MonthlyGoal monthlyGoal) {
 
 		return new MonthlyGoalListResponse(
 				monthlyGoal.getId(), monthlyGoal.getTitle(), monthlyGoal.isAchieved());
